+++ database/migrations/2024_06_08_062530_create_credit_notes_table.php
@@ -13,7 +13,7 @@ return new class extends Migration {
             DataBaseHElper::saleTable($table);
             //Datos solo de nota de credito
             $table->string('ncf_m',30)->nullable();
-            $table->float('n_available',4);
+            $table->decimal('n_available',15,4);
             $table->string('type',20)->default('Devolucion');
             $table->boolean('status')->default(true);
             $table->foreignIdFor(Sale::class, 'sale_id')
