+++ database/migrations/2024_09_18_104908_create_deleted_sales_table.php
@@ -12,10 +12,10 @@ return new class extends Migration {
             $table->id();
             $table->string('code',30)->unique();
             $table->foreignIdFor(Sale::class, 'sale_id');
-            $table->float('discount_amount',4)->default(0);
-            $table->float('tax',4);
-            $table->float('sub_total',4);
-            $table->float('amount',4);
+            $table->decimal('discount_amount',15,4)->default(0);
+            $table->decimal('tax',15,4);
+            $table->decimal('sub_total',15,4);
+            $table->decimal('amount',15,4);
             $table->boolean('status')->default(true);
             $table->boolean('close_table')->default(false);
             $table->softDeletes();
